requirements.txt
boto3==1.34.79
botocore==1.34.79
fairseq==0.12.2
GitPython==3.1.41
GitPython==3.1.43
google_api_python_client==2.116.0
llamaapi==0.1.36
matplotlib==3.8.2
mistralai==0.1.8
nltk==3.8.1
numpy==1.23.5
openai==1.16.2
pandas==2.2.1
pymongo==4.6.3
pytorch_lightning==2.1.3
qwikidata==0.4.2
Requests==2.31.0
rouge_score==0.1.2
sacrebleu==2.4.0
torch==2.1.2
transformers==4.37.1
Unidecode==1.3.8
urllib3==2.2.1
xmltodict==0.13.0
python-dotenv==1.0.1